trunk (#20)
* Fix stdafx include

* Fix basic handling of "Games" folder on Windows10 RS4 (#10)
This does the following:
- Sets the default state to hidden
- Skips the Games folder when searching

This does not:
- Hide the dead menu entry.

I do not currently know how to actively change the user preference setting to forcefully hide it.

* Add basic Visual Studio gitignore

* Add specific entries to gitignore

* Do not set default menu to Win7 on RS4 (#10)

* Rename "PC Settings" to "Settings" (#12)

* Create distinction between modern and legacy settings in search results

* Add more build artifacts to gitignore

* Add default paths for toolset and build all languages

* Fix several memsize, memtype and nullpointer issues

* create trunk branch containing all changes

* set fallback and next version to 4.3.2, set resource fallback value to allow loading in IDE

* add generated en-US.dll to gitignore

* Don't echo script contents, add disabled "git clean -dfx" to build fresh

* fix gitignore to not destroy en-US language project

+++ ClassicShellSrc/ClassicStartMenu/ClassicStartMenuDLL/ClassicStartMenuDLL.h
@@ -71,7 +71,7 @@ struct TaskbarInfo
 	bool HasPart( HWND part ) const;
 };
 
-TaskbarInfo *GetTaskbarInfo( int taskbarId );
+TaskbarInfo *GetTaskbarInfo( size_t taskbarId );
 UINT GetTaskbarPosition( HWND taskBar, MONITORINFO *pInfo, HMONITOR *pMonitor, RECT *pRc );
 
 extern HWND STARTMENUAPI g_TaskBar, g_OwnerWindow;
@@ -113,7 +113,7 @@ enum THotkeys
 // Set the hotkeys and controls for the start menu
 void EnableHotkeys( THotkeys enable );
 
-bool PointAroundStartButton( int taskbarId, const CPoint &pt=CPoint(GetMessagePos()) );
+bool PointAroundStartButton( size_t taskbarId, const CPoint &pt=CPoint(GetMessagePos()) );
 void ResetHotCorners( void );
 void RedrawTaskbars( void );
